@@ -1,7 +1,7 @@
 import {Component, ComponentRef, Input, Output, ViewChild, SimpleChange, EventEmitter} from "@angular/core";
 import {Router, RouteParams, RouterLink} from "@angular/router-deprecated";
 import {Observable} from "rxjs/Observable";
-import "rxjs/add/observable/forkjoin";
+import "rxjs/add/observable/forkJoin";
 
 import {ProductService, AccountService, VatTypeService} from "../../../../services/services";
 
@@ -240,8 +240,9 @@ export class ProductDetails {
            });
         });
     } 
-    
-    getComponentLayout(): ComponentLayout {
+
+    // TODO: return ComponentLayout when the object respects the interface
+    getComponentLayout(): any {
         return {
             Name: "Product",
             BaseEntity: "Product",
