@@ -33,8 +33,9 @@ export class EmailForm {
     }
  
     createFormConfig() {   
-        // TODO get it from the API and move these to backend migrations   
-        var view: ComponentLayout = {
+        // TODO get it from the API and move these to backend migrations
+        // TODO: turn to 'ComponentLayout when the object respects the interface
+        var view: any = {
             Name: "Email",
             BaseEntity: "Email",
             StatusCode: 0,
