@@ -34,8 +34,9 @@ export class PhoneForm {
     }
                
     createFormConfig() {   
-        // TODO get it from the API and move these to backend migrations   
-        var view: ComponentLayout = {
+        // TODO get it from the API and move these to backend migrations
+        // TODO: turn to 'ComponentLayout when the object respects the interface
+        var view: any = {
             StatusCode: 0,
             Name: "Phone",
             BaseEntity: "Phone",
