@@ -21,34 +21,34 @@ const CHILD_ROUTES = [
         useAsDefault: true,
         path: '/company',
         name: 'Firmainnstillinger',
-        loader: () => ComponentProxy.LoadComponentAsync('CompanySettings', './app/components/settings/companySettings/companySettings')
+        loader: () => ComponentProxy.LoadComponentAsync('CompanySettings','app/components/settings/companySettings/companySettings')
     }),
     new AsyncRoute({
         path: '/accounts',
         name: 'Kontoer',
-        loader: () => ComponentProxy.LoadComponentAsync('AccountSettings', './app/components/settings/accountSettings/accountSettings')
+        loader: () => ComponentProxy.LoadComponentAsync('AccountSettings','app/components/settings/accountSettings/accountSettings')
     }),
     new AsyncRoute({
         path: '/vat',
         name: 'VAT',
-        loader: () => ComponentProxy.LoadComponentAsync('VatSettings', './app/components/settings/vatSettings/vatSettings')
+        loader: () => ComponentProxy.LoadComponentAsync('VatSettings','app/components/settings/vatSettings/vatSettings')
     }),
     new AsyncRoute({
         path: '/user',
         name: 'Brukerinnstillinger',
-        loader: () => ComponentProxy.LoadComponentAsync('UserSettings', './app/components/settings/userSettings/userSettings')
+        loader: () => ComponentProxy.LoadComponentAsync('UserSettings','app/components/settings/userSettings/userSettings')
     }),
 
     new AsyncRoute({
         path: '/users',
         name: 'Brukere og roller',
-        loader: () => ComponentProxy.LoadComponentAsync('Users', './app/components/settings/users/users')
+        loader: () => ComponentProxy.LoadComponentAsync('Users','app/components/settings/users/users')
     }),
     
     new AsyncRoute({
         path: '/altinn',
         name: 'Altinn',
-        loader: () => ComponentProxy.LoadComponentAsync('AltinnSettings', './app/components/settings/altinnSettings/altinnSettings')
+        loader: () => ComponentProxy.LoadComponentAsync('AltinnSettings','app/components/settings/altinnSettings/altinnSettings')
     }),
 ];
 
